requirements.txt
# Core MCP dependencies
mcp>=1.0.0
mcp-server-fastmcp>=0.1.0

# Research and data processing
arxiv==2.1.0

# Environment and configuration
python-dotenv>=1.0.0

# LangChain and AI/LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0

# Async utilities
nest-asyncio>=1.5.0

# HTTP client (for MCP communications)
httpx>=0.25.0

# Type hints and utilities
typing-extensions>=4.0.0
